+++ src/avrtos/workqueue.c
@@ -179,13 +179,13 @@ void k_work_delayable_init(struct k_work_delayable *dwork, k_work_handler_t hand
 
 extern void z_event_schedule(struct k_event *event, k_timeout_t timeout);
 
-int k_work_delayable_schedule(struct k_workqueue *workqueue,
-							  struct k_work_delayable *dwork,
-							  k_timeout_t timeout)
+int8_t k_work_delayable_schedule(struct k_workqueue *workqueue,
+								 struct k_work_delayable *dwork,
+								 k_timeout_t timeout)
 {
 	Z_ARGS_CHECK(workqueue && dwork) return -EINVAL;
 
-	int ret			   = 0;
+	int8_t ret		   = 0;
 	const uint8_t lock = irq_lock();
 
 	/* Ensure the work item is not already pending for submission or already in
@@ -211,11 +211,11 @@ exit:
 	return ret;
 }
 
-int k_work_delayable_cancel(struct k_work_delayable *dwork)
+int8_t k_work_delayable_cancel(struct k_work_delayable *dwork)
 {
 	Z_ARGS_CHECK(workqueue && dwork) return -EINVAL;
 
-	int ret = k_event_cancel(&dwork->_event);
+	int8_t ret = k_event_cancel(&dwork->_event);
 	if (ret == -EAGAIN) {
 		const uint8_t lock = irq_lock();
 
@@ -234,7 +234,8 @@ int k_work_delayable_cancel(struct k_work_delayable *dwork)
 }
 
 #if CONFIG_SYSTEM_WORKQUEUE_ENABLE
-int k_system_work_delayable_schedule(struct k_work_delayable *dwork, k_timeout_t timeout)
+int8_t k_system_work_delayable_schedule(struct k_work_delayable *dwork,
+										k_timeout_t timeout)
 {
 	return k_work_delayable_schedule(&z_system_workqueue, dwork, timeout);
 }
